# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup 2026 StadiumAI — GenAI stadium operations platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AiQuery {
  module: string;
  query: string;
  /** @nullable */
  context?: string | null;
  /** @nullable */
  language?: string | null;
}

export interface AiResponse {
  answer: string;
  suggestions?: string[];
  module: string;
}

export interface IncidentTriageInput {
  description: string;
  location: string;
  /** @nullable */
  reportedBy?: string | null;
}

export interface IncidentTriageResult {
  priority: string;
  category: string;
  suggestedActions: string[];
  estimatedResolutionTime: string;
}

export interface CrowdInput {
  venue: string;
  zone: string;
  density: number;
  /** @nullable */
  matchPhase?: string | null;
}

export interface CrowdAnalysis {
  riskLevel: string;
  recommendations: string[];
  /** @nullable */
  alertMessage?: string | null;
}

export interface TranslationInput {
  text: string;
  targetLanguage: string;
  /** @nullable */
  sourceLanguage?: string | null;
}

export interface TranslationResult {
  translatedText: string;
  /** @nullable */
  detectedLanguage?: string | null;
}

export interface Incident {
  id: number;
  title: string;
  description: string;
  location: string;
  priority: string;
  status: string;
  module: string;
  createdAt: string;
  updatedAt: string;
}

export interface IncidentInput {
  title: string;
  description: string;
  location: string;
  priority: string;
  module: string;
}

export interface IncidentUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface Venue {
  id: number;
  name: string;
  city: string;
  country: string;
  capacity: number;
  currentOccupancy: number;
  crowdDensity: number;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  totalIncidents: number;
  openIncidents: number;
  criticalIncidents: number;
  resolvedToday: number;
  crowdCapacityPercent: number;
  activeVenues: number;
  volunteerCount: number;
  carbonSavedKg: number;
}

export interface Conversation {
  id: number;
  title: string;
  module: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
  module: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  module: string;
  createdAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export type ListIncidentsParams = {
status?: string;
priority?: string;
module?: string;
};