# lib/api-zod/src/generated/types/crowdAnalysis.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup 2026 StadiumAI — GenAI stadium operations platform
 * OpenAPI spec version: 0.1.0
 */

export interface CrowdAnalysis {
  riskLevel: string;
  recommendations: string[];
  /** @nullable */
  alertMessage?: string | null;
}