# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup 2026 StadiumAI — GenAI stadium operations platform
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: number;
  name: string;
  city: string;
  country: string;
  capacity: number;
  currentOccupancy: number;
  crowdDensity: number;
  status: string;
  createdAt: Date;
}